pyproject.toml
[tool.poetry]
name = "gitviper"
version = "0.1.0"
description = ""
authors = ["Jordan Hotmann <jhot@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests-toolbelt = "^1.0.0"
pydantic = "^2.8.2"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
result = "^0.17.0"


[tool.poetry.group.dev.dependencies]
datamodel-code-generator = "^0.25.9"
black = "^24.8.0"
types-requests = "^2.32.0.20240712"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
responses = "^0.25.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
